feat: move grouping of forms into selection element
