+++ Tests/Unit/Form/Element/FormcycleSelectionTest.php
@@ -1,0 +1,50 @@
+<?php
+
+namespace Xima\XmFormcycle\Tests\Unit\Form\Element;
+
+use PHPUnit\Framework\TestCase;
+use Xima\XmFormcycle\Form\Element\FormcycleSelection;
+
+class FormcycleSelectionTest extends TestCase
+{
+    public function testGroupForms()
+    {
+        self::assertEmpty(FormcycleSelection::groupForms([]));
+
+        $arr = [
+            [
+                'group' => 'b',
+                'name' => 'Foo',
+            ],
+            [
+                'name' => 'Bar',
+            ],
+            [
+                'group' => 'a',
+                'name' => 'Hossa',
+            ],
+        ];
+
+        $result = FormcycleSelection::groupForms($arr);
+
+        self::assertEquals([
+            'a' => [
+                [
+                    'group' => 'a',
+                    'name' => 'Hossa',
+                ],
+            ],
+            'b' => [
+                [
+                    'group' => 'b',
+                    'name' => 'Foo',
+                ],
+            ],
+            0 => [
+                [
+                    'name' => 'Bar',
+                ],
+            ],
+        ], $result);
+    }
+}
